requirements.txt
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.6
pem==21.1.0
pycparser==2.20
pyOpenSSL==20.0.1
six==1.15.0
